requirements.txt
pandas
promptflow
promptflow-tools
azure-identity
promptflow-evals